Finish the QSO Window and general bug fixes
As well as finishing the QSO Window implementation I have added logic to remember the windows sizes and position so they are reopened with the size and location of what they were before. Also there were quite a few bugs in the logic of the main window which have been fixed.